refactor: modify code
