+++ src/components/link/LinkViewComponent.tsx
